server/requirements.txt
flask
flask-cors
gunicorn
google-genai
youtube-transcript-api
python-dotenv